sdk/requirements.txt
nats-py>=2.9.0
protobuf>=5.29.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0

